DemoApp/backend/requirements.txt
flask
python-dotenv
Pillow
flask-cors